Abstract timeout functions and use on alert detailspage
